requirements.txt
absl-py==2.1.0
ace_tools==0.0
asttokens==2.0.5
astunparse==1.6.3
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.5.0
certifi==2022.6.15
cffi==1.17.0
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==4.4.2
entrypoints==0.4
executing==0.8.3
facenet-pytorch==2.5.2
Far==0.0.1
fer==22.5.1
ffmpeg==1.4
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.33.3
fsspec==2023.12.2
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
huggingface-hub==0.20.1
idna==3.3
imageio==2.35.1
imageio-ffmpeg==0.5.1
ipykernel==6.13.0
ipython==8.3.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.18.1
Jinja2==3.1.4
jupyter-client==7.3.1
jupyter-core==4.10.0
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.5.5
networkx==3.3
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
proglog==0.1.10
prompt-toolkit==3.0.29
protobuf==4.25.3
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.22
pygame==2.6.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pynput==1.7.7
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.8.2
pytweening==1.2.0
pytz==2022.1
pywin32==304
PyYAML==6.0.1
pyzmq==22.3.0
requests==2.28.0
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
scipy==1.14.0
six==1.16.0
sounddevice==0.4.7
soupsieve==2.3.2.post1
stack-data==0.2.0
sympy==1.14.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.1
tqdm==4.66.1
traitlets==5.2.0
typing_extensions==4.13.2
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==3.0.4
wrapt==1.14.1
